Custom index file and test improvement
Add support for node-static custom index files. This is in not-yet released version of node-static, so using the node-static GitHub repo instead of npm package.

Slight rewrite of the plugin and new `shutdown()` helper method. This is intended to use when testing, in order to avoid having to use multiple port numbers between test suites.

Added new test cases.